Don't overwrite JSON files if they can't be parsed Based on recent events of people having their bookmarks / watched replays file being removed, I thought of adding some logic to avoid (or at least decrease) such occurrences. Even if a JSON file is not parseable, it doesn't give us the right to overwrite it, as the user might have manually edited them to do some testing and _might_ have messed up the syntax. I'm definetely not talking about me :)
Anyway, it works by simply catching the exception thrown by JSON.parse()
and then it tells the user the program is unable to continue because
some JSON file does not have a proper syntax.

Then it gives the user three options:

* Ask for help in our Discord group
* Reset the file (with a big-ass warning)
* Close the dialog

The 'Discord' button will simply open our Discord group URL in their
browser.

The 'Reset' button will open a new dialog asking for user confirmation.
If the user clicks the 'Yes' button, then the JSON file shall be
overwritten by an empty array -> [].
If the user clicks the 'No' button, it will go back to the previous
dialog.

And finally, the 'Close' button will close the dialog and terminate the
program without overwriting anything.