Fix AppImage icons on Linux
It turns out that if you specify an absolute path for the "icon"
property, the icon will be (wrongfully) installed in:

~/.local/share/icons/hicolor/0x0/apps/appimagekit-liveme-pro-tools.png

(notice the 0x0) there. That's not a valid size for an icon.

So the fix is to set the "icon" property pointing to the directory
containing all icons. Then electron-builder will deal with the rest.