Better error message when searching for videos
We can't be sure if the replay is still being generated or was deleted.
You can observe this by waiting on someone's live to go offline, then
you immediately grab the video ID and try looking it up. Both
`videosource` and `hlsvideosource` properties are going to be empty
until the replay is fully generated.