Add new player config
This is one of the features I was missing the most: persistent player
configuration (such as volume and mute state). It was really frustrating
having your volume set to 100% every time you'd watch a stream. Not
anymore!

For now it only saves the volume and mute states, but there is nothing
preventing other customizations from being added in the future.