Update pickProperSource() / getVideoEndDate()
I've a parameter called preferLiveFlv, so when it's set to true,
pickProperSource() will try to return the URL of the FLV stream, instead
of the HLS one. This allow us for smoother playback when dealing with
live streams.

This options is off by default, and as of now it only affects live streams
in the internal video player.